Add files via upload
This contains information on RNA signals for MS patients. I am specifically looking at DRB5 Signal (this is a gene that people either have or don't). DRB1 and DRA signals are used as controls because everyone should express these. I used a model in PyTorch with the DRB5, DRB1, and DRA signals to predict a patient's diagnosis.